Fix type cast

+++ src/driver/cpp17_mock.hpp
@@ -20,7 +20,7 @@ auto randomVariation(uint16_t current) -> uint16_t {
     std::mt19937 rng(dev());
     std::uniform_int_distribution<std::mt19937::result_type> dist(0, 16);
     if (current == 1024) return current - dist(rng);
-    return std::min(current + dist(rng) - 15, 1024);
+    return std::min(current + static_cast<uint16_t>(dist(rng)) - 15, 1024);
 }
 
 SoilResistivity::SoilResistivity(const iop_hal::PinRaw powerPin) noexcept: sensor(new uint16_t(randomValue())) { (void) powerPin; }
